refactor: change logic of using mock data
