+++ product-service/README.md
@@ -1,5 +1,7 @@
 # Serverless - AWS Node.js Typescript
 
+Product service API build with AWS API Gateway and AWS Lambda.
+
 This project has been generated using the `aws-nodejs-typescript` template from the [Serverless framework](https://www.serverless.com/).
 
 For detailed instructions, please refer to the [documentation](https://www.serverless.com/framework/docs/providers/aws/).
