+++ product-service/src/functions/__tests__/get-products-by-id.spec.ts
@@ -1,5 +1,5 @@
 import { getProductByIsbn } from '../get-product-by-id/handler';
-import data from '../../data/books.json';
+import data from '../../mock-data/books.json';
 
 describe('Lambda getProductsById function', () => {
   test('returns the correct product from the list', async () => {
